requirements.txt
pandas>=2.2.1
tqdm>=4.66.2
opencv-python>=4.9.0.80
numpy>=1.26.4
pillow>=10.2.0
pytz>=2024.1
imgaug>=0.4.0
scikit-image>=0.22.0
shapely>=2.0.3
matplotlib>=3.8.3